RFC: feat(livenet): add memory size check depending on live image size
For writeable live images, the memory must hold the complete image.
That means it must be at least the size of the image plus some RAM
necessary to run processes.
With too small RAM, the OOM Killer steps in and makes the boot hang.
This patch lets the system go into the emergency shell instead.

The minimum RAM of 512Mb is just arbitrarily choosen.
We could set any other number that appears appropriate instead.
Alternatively, we could add a new parameter rd.min.ram or such.